+++ src/routes/Router.tsx
@@ -20,7 +20,6 @@ import NewsPage from "../pages/news/NewsPage";
 import Base from "../Base";
 import MyPagePage from "~/pages/MyPagePage";
 import SignUpPage from "~/pages/SignUpPage";
-import LoginPage from "~/pages/LoginPage";
 
 const Router = () => {
   return (
@@ -31,7 +30,6 @@ const Router = () => {
 
           <Route path="/my_page" element={<MyPagePage />} />
           <Route path="/signup" element={<SignUpPage />} />
-          <Route path="/login" element={<LoginPage />} />
 
           <Route path="/club/about" element={<AboutPage />} />
           <Route path="/club/organization" element={<OrganizationPage />} />
